refactor(web): use axios to talk to the questions service
